refactor: HandlerMapping 추상화
