+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handler/HandlerMapping.java
@@ -1,0 +1,10 @@
+package com.interface21.webmvc.servlet.mvc.tobe.handler;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+public interface HandlerMapping {
+
+    void initialize();
+
+    HandlerExecution getHandler(HttpServletRequest request);
+}
